Reorganize files in preparation to resolve merge conflicts
